Add nav dropdown animation when clicked | Fix style
