+++ packages/client/src/index.css
@@ -20,6 +20,7 @@ html, body {
   height: 100%;
   font-family: 'Apercu Light', sans-serif;
   background-color: #fff;
+  overflow: hidden;
 }
 
 body {
